chore: use explicit type for defaults (#729)
Fixes the [default_trait_access](https://rust-lang.github.io/rust-clippy/master/index.html#default_trait_access) and explicitly declares a few lints as allowed.  This makes it a bit cleaner which type is where.

One thing I noticed, but didn't change yet, is that in some cases it might be easier to write `..Default::default()` at the end of each object, and get rid of `id` and other fields with defaults?